requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.33.0
APScheduler==3.10.4
